Reuse existing token if possible and also attempt login at max once on failures.
The idea is to reuse the token (thread-safe) if it has not expired, and
also automatically re-auth if we fail to query the cable modem (provided
we didn't just login prior to the query).

Every query also extends the expiry of the token automatically by a
further 10 mins.

This number could potentially be increased, but from the testing this
seems to be a safe number without any issues.