Add more informative logging/errors when dealing with 404 errors with the query action.
If the query action runs into a 404 (and we have logic to ensure the
credentials have not expired before attempting a query action), it means
the host might have some reset and hence erased all stored sessions.

Hence we log this case more explicitly now.

In the future, it is better to convert this to an explicit error type
that can be checked by the caller and handled better.