feat: Remove EstablishedAt field.
EstablishedAt is a derived field and can be computed using the
combination of SystemTime and UpTime fields instead.

BREAKING CHANGE:

This is technically a breaking change since a field that is being
exported is being removed. Although, this is still not a stable version
and hence it is fine for users to adapt to this change.